fix AI::StrongerEnemyArmy
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@2049 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/army/army.h
@@ -128,7 +128,7 @@ namespace Army
 	    u32		GetHitPoints(void) const;
 	    u32		GetDamageMin(void) const;
 	    u32		GetDamageMax(void) const;
-	    double	GetStrength(void) const;
+	    u32		GetStrength(void) const;
 
 	    bool	isValid(void) const;
 	    bool	HasMonster(const Monster &) const;
